Added outside listener to date dialog to listen for date changes.
Note that the listener is explicitly registered with the "okay"
button; we need to ensure a specific order of execution.

Added date listener class to even dialog to change the date text
fields when a new date is selected.

The date text fields are now instance variables since they
get updated from the model similar to the other text fields.