Fixed bug in ReminderDialog and added Event rendering.
Does not work if user does not go into the dialog.
Fixed by creating a default Reminder with a default
Date and Time set.

Also fixed Reminder class equals method which compared
references rather than using the equals methods of
Date and Time.

Reformatted the calendar saving/load code.

Added Map to the month table's model to render sorted set's of
Events. Renders events on dates from their start date to their
end date inclusive. Uses default sorted set renderer (toString).

Date increment/decrements method now returns a new Date that is
one day ahead/behind of the date object the operation was operated
on. This is necessary because the behaviour of the Map structure is
not defined when a key is modified. Month increment/decrement
parallels have also been provided for the next/previous buttons
in the month view; day incrementing is now too slow to
perform the month increment because of the object creation.